refactor!: extract Emulator enum to shared constants, fix ESFMU→ESFMu
