+++ src/libadlmidi.js
@@ -1,19 +1,22 @@
 /**
  * libADLMIDI-JS - Main Thread Interface
- * 
+ *
  * High-level API for real-time OPL3 FM synthesis in the browser.
- * 
+ *
  * @example
  * ```javascript
  * import { AdlMidi } from 'libadlmidi-js';
- * 
+ *
  * const synth = new AdlMidi();
  * await synth.init('/path/to/processor.js');
- * 
+ *
  * synth.noteOn(0, 60, 100);  // Middle C on channel 0
  * synth.noteOff(0, 60);
  * ```
  */
+
+import { Emulator } from './utils/constants.js';
+export { Emulator };
 
 /**
  * Bank identifier for instrument access
@@ -72,40 +75,6 @@
  * @property {boolean} [deepVibrato] - Enable deep vibrato
  * @property {boolean} [deepTremolo] - Enable deep tremolo
  */
-
-/**
- * Available OPL2/OPL3 emulator cores.
- * Use with switchEmulator() to change the synthesis engine at runtime.
- * Note: Only emulators compiled into the current profile are available.
- * @readonly
- * @enum {number}
- */
-export const Emulator = Object.freeze({
-    /** Nuked OPL3 v1.8 - Most accurate, higher CPU usage */
-    NUKED: 0,
-    /** Nuked OPL3 v1.7.4 - Slightly older version */
-    NUKED_174: 1,
-    /** DosBox OPL3 - Good accuracy, lower CPU usage */
-    DOSBOX: 2,
-    /** Opal - Reality Adlib Tracker emulator */
-    OPAL: 3,
-    /** Java OPL3 - Port of emu8950 */
-    JAVA: 4,
-    /** ESFMu - ESFM chip emulator */
-    ESFMU: 5,
-    /** MAME OPL2 */
-    MAME_OPL2: 6,
-    /** YMFM OPL2 */
-    YMFM_OPL2: 7,
-    /** YMFM OPL3 */
-    YMFM_OPL3: 8,
-    /** Nuked OPL2 LLE - Transistor-level emulation */
-    NUKED_OPL2_LLE: 9,
-    /** Nuked OPL3 LLE - Transistor-level emulation */
-    NUKED_OPL3_LLE: 10,
-    /** Nuked OPL2 Lite - Lightweight OPL2 emulation for AdLib-era music */
-    NUKED_OPL2_LITE: 11,
-});
 
 export class AdlMidi {
     /** @type {boolean} */
